Move NEWOBJ regression to serialization tests
