+++ test/distributed/tensor/test_placement_types.py
@@ -1,10 +1,6 @@
 # Owner(s): ["oncall: distributed"]
 import copy
-import io
 import itertools
-import pickle
-import sys
-import zipfile
 
 import sympy
 
@@ -33,39 +29,6 @@ from torch.testing._internal.common_utils import run_tests, TestCase
 
 # Basic functionality test for Placement types.
 class PlacementTypesTestCase(TestCase):
-    # SECURITY: This prevents malformed pickles from exposing uninitialized
-    # pybind11 objects. Do not remove this regression test.
-    def test_weights_only_newobj_requires_build(self):
-        malformed_pickle = (
-            pickle.PROTO
-            + b"\x02"
-            + pickle.GLOBAL
-            + Shard.__module__.encode()
-            + b"\n"
-            + Shard.__name__.encode()
-            + b"\n"
-            + pickle.EMPTY_TUPLE
-            + pickle.NEWOBJ
-            + pickle.STOP
-        )
-
-        checkpoint = io.BytesIO()
-        with zipfile.ZipFile(checkpoint, "w") as archive:
-            archive.writestr("archive/data.pkl", malformed_pickle)
-            archive.writestr("archive/version", "3\n")
-            archive.writestr("archive/byteorder", sys.byteorder)
-        checkpoint.seek(0)
-
-        with self.assertRaisesRegex(
-            pickle.UnpicklingError, "pickle data is likely corrupt or malicious"
-        ):
-            torch.load(checkpoint, weights_only=True)
-
-        buffer = io.BytesIO()
-        torch.save(Shard(2), buffer)
-        buffer.seek(0)
-        self.assertEqual(torch.load(buffer, weights_only=True), Shard(2))
-
     def test_type_identification(self):
         shard = Shard(3)
         strided_shard = _StridedShard(dim=3, split_factor=7)
